web_transbank_1/BACKEND/project/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-allauth==0.44.0
django-appconf==1.0.4
django-cors-headers==3.7.0
django-cryptography==1.0
django-fernet-fields==0.6
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
geographiclib==1.52
geopy==2.1.0
idna==2.10
marshmallow==2.15.6
oauthlib==3.1.0
Pillow==8.2.0
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.9.0
sqlparse==0.4.1
transbank-sdk==1.4.0
urllib3==1.26.4
